requirements.txt
opencv-python-headless
numpy
mediapipe
PyYAML
